requirements.txt
dj-database-url==0.4.0
Django==1.10
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
django-emailit>=0.2.2
django-bootstrap3
django-extensions==1.7.2
markdown
django-localflavor
pyyaml
ShopifyAPI
django-bootstrap3-datetimepicker
celery
django-celery
